creates the class StandardController which implements ClientMissionController
